requirements.txt
discord.py
discord-ext-audiorec
openai
pydub
python-dotenv
